+++ Tests/Mock/DtoClass.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace Wakeapp\Component\DtoResolver\Tests\Mock;
 
+use ArrayAccess;
+use Wakeapp\Component\DtoResolver\Dto\DtoArrayAccessTrait;
 use Wakeapp\Component\DtoResolver\Dto\DtoResolverInterface;
 use Wakeapp\Component\DtoResolver\Dto\DtoResolverTrait;
 
-class DtoClass implements DtoResolverInterface
+class DtoClass implements DtoResolverInterface, ArrayAccess
 {
     use DtoResolverTrait;
+    use DtoArrayAccessTrait;
 
     public $publicProperty;
     protected $protectedProperty;